Changes from first Linux Port attempt.  Needed typecast on QByteArray indexes.  Fixed the memory leaks found via valgrind.  Fixed KJVPhraseEditor menu issues (but looks like we have some more if you quickly try to delete the phrases while they are still updating... hmmmm).  UI tweaks and font settings to get the windows layout to appear consistent between the two platforms.  Added "Please Wait" message to the splash screen.  Main Window icon setting, so it's set in Linux as well.  Switched the PassageNavigator ToolTip to use our CToolTipEdit.  Added some code to import CToolTipEdit color palette, color roles, and window sizing, but still have issues with the color in general not following the system styles.  It seems they are doing some monkey business with the QLabel object.  Added namespace to remaining connect statements in CScriptureEdit -- for some strange reason it compiled OK as it was both on Windows and on Fedora and even Ubuntu netbook, but not on Ubuntu desktop.  Adding the namespace I guess disambiguates it and works either way -- also seems to work regardless of using T:: or U::.  Picked T:: for consistency.
git-svn-id: file:///home/dewhisna/Documents/programs/Bible/svndump/purebiblesearch_150731/code_gitcvt/trunk@198 523e89ec-37fb-4b40-b97f-42a194c95123

+++ programs/KJVCanOpener/BuildDB.cpp
@@ -493,12 +493,12 @@ QByteArray CBuildDatabase::CSVStringToIndexBlob(const QString &str)
 		nValue = slValues.at(i).toUInt();
 		assert(nValue != 0);
 		for (unsigned int j=1; j<=sizeof(uint32_t); ++j) {
-			baBlob[(i*sizeof(uint32_t))+(sizeof(uint32_t)-j)] = (nValue & 0xFF);
+			baBlob[static_cast<unsigned int>((i*sizeof(uint32_t))+(sizeof(uint32_t)-j))] = (nValue & 0xFF);
 			nValue = nValue >> 8;
 		}
 	}
 	assert((baBlob.size() % sizeof(uint32_t)) == 0);
-	assert(baBlob.size() == (slValues.size() * sizeof(uint32_t)));
+	assert(baBlob.size() == static_cast<int>(slValues.size() * sizeof(uint32_t)));
 
 	return baBlob;
 }
